+++ java_wrong_method_type_exception_test.go
@@ -12,15 +12,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package exception
+package hessian
 
 import (
 	"github.com/stretchr/testify/assert"
 	"testing"
 )
 
-func TestMalformedParameterizedTypeException(t *testing.T) {
-	testDecodeFrameworkFunc(t, "throw_MalformedParameterizedTypeException", func(r interface{}) {
-		assert.Equal(t, "MalformedParameterizedType", r.(error).Error())
+func TestWrongMethodTypeException(t *testing.T) {
+	testDecodeFrameworkFunc(t, "throw_WrongMethodTypeException", func(r interface{}) {
+		assert.Equal(t, "WrongMethodTypeException", r.(error).Error())
 	})
 }
